Restructure entire repo
