+++ cs-serviceclient-sample/.gitignore
