+++ cs-serviceclient-sample/cs-serviceclient-sample.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cs-serviceclient-sample", "cs-serviceclient-sample\cs-serviceclient-sample.csproj", "{14C7ABD2-8FBA-4D24-BA58-5848FFDBABB0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cs-serviceclient", "..\cs-serviceclient\cs-serviceclient\cs-serviceclient.csproj", "{0014303C-5C5C-4C34-8406-4FE30E6BDBE8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{14C7ABD2-8FBA-4D24-BA58-5848FFDBABB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14C7ABD2-8FBA-4D24-BA58-5848FFDBABB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14C7ABD2-8FBA-4D24-BA58-5848FFDBABB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14C7ABD2-8FBA-4D24-BA58-5848FFDBABB0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0014303C-5C5C-4C34-8406-4FE30E6BDBE8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0014303C-5C5C-4C34-8406-4FE30E6BDBE8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0014303C-5C5C-4C34-8406-4FE30E6BDBE8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0014303C-5C5C-4C34-8406-4FE30E6BDBE8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
