+++ cs-serviceclient-sample/cs-serviceclient-sample.sln.DotSettings
