+++ cs-serviceclient-sample/cs-serviceclient-sample/Annotations.cs
