+++ cs-serviceclient-sample/cs-serviceclient-sample/App.config
