+++ cs-serviceclient-sample/cs-serviceclient-sample/App.xaml.cs
