+++ cs-serviceclient-sample/cs-serviceclient-sample/AssemblyInfo.cs
