+++ cs-serviceclient-sample/cs-serviceclient-sample/Core/ObservableObject.cs
