+++ cs-serviceclient-sample/cs-serviceclient-sample/MVVM/Model/FPlusCommunicator.cs
