+++ cs-serviceclient-sample/cs-serviceclient-sample/MVVM/View/HomeView.xaml
