+++ cs-serviceclient-sample/cs-serviceclient-sample/MVVM/View/SettingsView.xaml
