+++ cs-serviceclient-sample/cs-serviceclient-sample/MVVM/ViewModel/HomeViewModel.cs
