+++ cs-serviceclient-sample/cs-serviceclient-sample/MVVM/ViewModel/MainViewModel.cs
