+++ cs-serviceclient-sample/cs-serviceclient-sample/MVVM/ViewModel/SettingsViewModel.cs
