+++ cs-serviceclient-sample/cs-serviceclient-sample/MainWindow.xaml.cs
