+++ cs-serviceclient-sample/cs-serviceclient-sample/Properties/Settings.settings
