+++ cs-serviceclient-sample/cs-serviceclient-sample/Theme/MenuButtonTheme.xaml
