+++ cs-serviceclient/cs-serviceclient/cs-serviceclient.csproj
